update the folder strucutre
