+++ app/page.jsx
@@ -1,13 +1,13 @@
 // import Modal from "@/components/home/Modal/Modal";
-import Banner from "@/components/home_comp/Banner/Banner";
-import InvestSolutions from "@/components/home_comp/InvestSolutions";
-import LiveData from "@/components/home_comp/LiveData";
-import NewsLetter from "@/components/home_comp/NewsLetter";
-import Services from "@/components/home_comp/Services";
-import StartTrading from "@/components/home_comp/StartTrading";
-import Subscription from "@/components/home_comp/Subscription";
-import TradingStep from "@/components/home_comp/TradingStep";
-import WhyChooseUs from "@/components/home_comp/WhyChooseUs";
+import Banner from "@/components/root_comp/home_comp/Banner/Banner";
+import InvestSolutions from "@/components/root_comp/home_comp/InvestSolutions";
+import LiveData from "@/components/root_comp/home_comp/LiveData";
+import NewsLetter from "@/components/root_comp/home_comp/NewsLetter";
+import Services from "@/components/root_comp/home_comp/Services";
+import StartTrading from "@/components/root_comp/home_comp/StartTrading";
+import Subscription from "@/components/root_comp/home_comp/Subscription";
+import TradingStep from "@/components/root_comp/home_comp/TradingStep";
+import WhyChooseUs from "@/components/root_comp/home_comp/WhyChooseUs";
 
 export default function Home() {
   return (
@@ -21,7 +21,6 @@ export default function Home() {
       <LiveData />
       <Subscription />
       <NewsLetter />
-      {/* <Modal /> */}
     </main>
   );
 }
