+++ app/register/page.jsx
@@ -12,10 +12,10 @@ import ReCAPTCHA from "react-google-recaptcha";
 import { useRouter } from "next/navigation";
 import usePublicAPI from "@/hooks/usePublicAPI";
 import Lottie from "lottie-react";
-import ArrowBackIcon from '@mui/icons-material/ArrowBack';
-import RegisterAnim from "../../assets/regiAnim"
+import ArrowBackIcon from "@mui/icons-material/ArrowBack";
+import RegisterAnim from "../../assets/regiAnim";
 import Magnetic from "@/components/library/Magnetic";
-import SocialLogin from "@/components/auth_comp/SocialLogin";
+import SocialLogin from "@/components/root_comp/auth_comp/SocialLogin";
 
 // customized TextField
 const CssTextField = styled(TextField)({
@@ -99,19 +99,34 @@ const Register = () => {
       });
   };
 
-
   return (
     <div className="flex flex-col xl:flex-row xl:items-center min-h-[100vh] relative">
       <Magnetic>
-        <Link href="/" className="text-white font-semibold flex items-center gap-3 absolute top-5 2xl:top-10 left-7 2xl:left-12 z-10"><ArrowBackIcon />Home</Link>
+        <Link
+          href="/"
+          className="text-white font-semibold flex items-center gap-3 absolute top-5 2xl:top-10 left-7 2xl:left-12 z-10"
+        >
+          <ArrowBackIcon />
+          Home
+        </Link>
       </Magnetic>
-      <Stack flex={1} sx={{ backgroundColor: "#455ce9", height: "100vh", display: "flex", justifyContent: "center", alignItems: "center" }}>
-        <Lottie className=" w-3/4" animationData={RegisterAnim} loop={true} />
-      </Stack>
       <Stack
         flex={1}
+        sx={{
+          backgroundColor: "#455ce9",
+          height: "100vh",
+          display: "flex",
+          justifyContent: "center",
+          alignItems: "center",
+        }}
       >
-        <form onSubmit={handleSubmit(onSubmit)} className="px-5 mx-auto my-10 xl:my-0 md:px-0 w-full md:w-[80%] 2xl:w-[70%]">
+        <Lottie className=" w-3/4" animationData={RegisterAnim} loop={true} />
+      </Stack>
+      <Stack flex={1}>
+        <form
+          onSubmit={handleSubmit(onSubmit)}
+          className="px-5 mx-auto my-10 xl:my-0 md:px-0 w-full md:w-[80%] 2xl:w-[70%]"
+        >
           <Typography
             variant="h2"
             mb={2}
@@ -121,7 +136,6 @@ const Register = () => {
           >
             Create Account
           </Typography>
-
 
           <Stack mt={4} gap={3}>
             {/* user name */}
@@ -205,7 +219,10 @@ const Register = () => {
 
             <Stack mt={2} alignItems="center">
               {captchaValue ? (
-                <Button className="w-full" type="submit"> Create Account</Button>
+                <Button className="w-full" type="submit">
+                  {" "}
+                  Create Account
+                </Button>
               ) : (
                 <Button disabled className="w-full" type="submit">
                   Create Account
@@ -214,7 +231,10 @@ const Register = () => {
             </Stack>
             <Typography className="text-lg lg:text-xl text-center">
               Have an account?{" "}
-              <Link href="/login" className=" font-bold text-primary hover:underline">
+              <Link
+                href="/login"
+                className=" font-bold text-primary hover:underline"
+              >
                 Login here
               </Link>
             </Typography>
