+++ components/common/root_nav/root_nav_comp/footer/index.css
