+++ components/layout/RootLayout.jsx
@@ -1,8 +1,8 @@
 "use client";
 import React from "react";
-import Footer from "../footer/Footer";
+import Footer from "../common/footer/Footer";
 import { usePathname } from "next/navigation";
-import RootNav from "../navigation/root_nav/RootNav";
+import RootNav from "../common/root_nav/RootNav";
 
 const MainLayout = ({ children }) => {
   const pathname = usePathname();
