+++ components/root_comp/home_comp/Services.jsx
@@ -1,24 +1,17 @@
 "use client";
-
 import React from "react";
-import Card from "@mui/material/Card";
 import CardContent from "@mui/material/CardContent";
 import Typography from "@mui/material/Typography";
 import Image from "next/image";
-import Icon1 from "../../assets/Services/icon-1.png";
-import Icon2 from "../../assets/Services/icon-2.png";
-import Icon3 from "../../assets/Services/icon-3.png";
-import Icon4 from "../../assets/Services/icon-4.png";
-import Icon5 from "../../assets/Services/icon-5.png";
-import Icon6 from "../../assets/Services/icon-6.png";
-
+import Icon1 from "../../../assets/Services/icon-1.png";
+import Icon2 from "../../../assets/Services/icon-2.png";
+import Icon3 from "../../../assets/Services/icon-3.png";
+import Icon4 from "../../../assets/Services/icon-4.png";
+import Icon5 from "../../../assets/Services/icon-5.png";
+import Icon6 from "../../../assets/Services/icon-6.png";
 import Container from "@/components/library/Container";
-
-//framer motion
 import { motion } from "framer-motion";
-
-// variants
-import { fadeIn } from "../utils/variants";
+import { fadeIn } from "../../utils/variants";
 
 const Services = () => {
   return (
