+++ components/root_comp/home_comp/Subscription.jsx
@@ -1,8 +1,8 @@
-'use client'
+"use client";
 import Image from "next/image";
 
 //import image from assets file
-import logo from "../../assets/nextrade-logo.png";
+import logo from "../../../assets/nextrade-logo.png";
 import Container from "@/components/library/Container";
 
 // material icon import
@@ -13,7 +13,7 @@ import Button from "@/components/library/buttons/root_button/RootButton";
 import { motion } from "framer-motion";
 
 // variants
-import { fadeIn } from "../utils/variants";
+import { fadeIn } from "../../utils/variants";
 
 const Subscription = () => {
   return (
@@ -22,30 +22,33 @@ const Subscription = () => {
         <div className=" text-center">
           <Image src={logo} alt="logo" width={80} className="my-2 mx-auto" />
           <motion.h1
-          variants={fadeIn("up", 0.2)}
-          initial="hidden"
-          whileInView={"show"}
-          viewport={{ once: false, amount: 0.1 }}
-          className="lg:text-3xl md:text-3xl text-4xl my-4 text-primary font-bold">
+            variants={fadeIn("up", 0.2)}
+            initial="hidden"
+            whileInView={"show"}
+            viewport={{ once: false, amount: 0.1 }}
+            className="lg:text-3xl md:text-3xl text-4xl my-4 text-primary font-bold"
+          >
             Get More Benefits With Subscription
           </motion.h1>
           <motion.p
-          variants={fadeIn("up", 0.4)}
-          initial="hidden"
-          whileInView={"show"}
-          viewport={{ once: false, amount: 0.1 }}
-          className="text-lg my-4 text-slate-500 font-medium">
+            variants={fadeIn("up", 0.4)}
+            initial="hidden"
+            whileInView={"show"}
+            viewport={{ once: false, amount: 0.1 }}
+            className="text-lg my-4 text-slate-500 font-medium"
+          >
             Monthly and yearly Subscription Sell
           </motion.p>
         </div>
         <div className=" grid xl:grid-cols-3 lg:grid-cols-2 grid-cols-1 gap-10 mt-16">
           {/* cart 1 basic plan */}
           <motion.div
-          variants={fadeIn("right", 0.5)}
-      initial="hidden"
-      whileInView={"show"}
-      viewport={{ once: false, amount: 0.1 }}
-          className=" border border-gray-200  shadow p-6 py-8  rounded-lg ">
+            variants={fadeIn("right", 0.5)}
+            initial="hidden"
+            whileInView={"show"}
+            viewport={{ once: false, amount: 0.1 }}
+            className=" border border-gray-200  shadow p-6 py-8  rounded-lg "
+          >
             <h3 className=" text-xl font-bold text-gray-600">Basic Plan</h3>
             <h1 className=" text-3xl font-extrabold py-3 ">
               $99.99{" "}
@@ -93,11 +96,12 @@ const Subscription = () => {
           </motion.div>
           {/* cart 2 Standard plan */}
           <motion.div
-          variants={fadeIn("up", 0.6)}
-      initial="hidden"
-      whileInView={"show"}
-      viewport={{ once: false, amount: 0.1 }}
-          className=" border  border-gray-200 shadow p-6 py-8  rounded-lg ">
+            variants={fadeIn("up", 0.6)}
+            initial="hidden"
+            whileInView={"show"}
+            viewport={{ once: false, amount: 0.1 }}
+            className=" border  border-gray-200 shadow p-6 py-8  rounded-lg "
+          >
             <h3 className=" text-xl font-bold text-gray-600">Standard Plan</h3>
             <h1 className=" text-3xl font-extrabold py-3 ">
               $800{" "}
@@ -150,11 +154,12 @@ const Subscription = () => {
           </motion.div>
           {/* cart 1 Premium  plan */}
           <motion.div
-          variants={fadeIn("left", 0.7)}
-      initial="hidden"
-      whileInView={"show"}
-      viewport={{ once: false, amount: 0.1 }}
-          className=" border border-gray-200 shadow p-6 py-8  rounded-lg ">
+            variants={fadeIn("left", 0.7)}
+            initial="hidden"
+            whileInView={"show"}
+            viewport={{ once: false, amount: 0.1 }}
+            className=" border border-gray-200 shadow p-6 py-8  rounded-lg "
+          >
             <h3 className=" text-xl font-bold text-gray-600">Premium Plan</h3>
             <h1 className=" text-3xl font-extrabold py-3 ">
               $1200{" "}
