+++ components/root_comp/home_comp/liveCoinPrice.jsx
@@ -1,0 +1,26 @@
+"use client";
+import Container from "../../library/Container";
+import { TickerTape } from "react-ts-tradingview-widgets";
+//framer motion
+import { motion } from "framer-motion";
+
+// variants
+import { fadeIn } from "../../utils/variants";
+
+const liveCoinPrice = () => {
+  return (
+    <Container>
+      <motion.div
+        variants={fadeIn("down", 0.2)}
+        initial="hidden"
+        whileInView={"show"}
+        viewport={{ once: false, amount: 0.1 }}
+        className="absolute bottom-5 w-full"
+      >
+        <TickerTape displayMode="adaptive"></TickerTape>
+      </motion.div>
+    </Container>
+  );
+};
+
+export default liveCoinPrice;
